requirements.txt
amqp==2.1.4
billiard==3.5.0.2
boto3==1.4.4
botocore==1.5.55
celery==4.0.2
Django==1.11.1
django-celery-results==1.0.1
django-filter==1.0.4
djangorestframework==3.6.3
docutils==0.13.1
jmespath==0.9.2
kombu==4.0.2
Markdown==2.6.8
mysqlclient==1.3.10
python-dateutil==2.6.0
python-memcached==1.58
pytz==2017.2
redis==2.10.5
s3transfer==0.1.10
six==1.10.0
vine==1.1.3
